Revert "add: fix tests to pass build"
